Autogen First Commit for PHP
Summary:
We are introducing the autogen SDK for Python and this diff will be the first commit.
Now the core classes will be autogenerated from live API definitions,
which will cut down the majority of SDK bugs, especially the ones related to
missing endpoints or fields.

We won't accept pull requests for these autogenerated core classes, which will
have a warning at the header, and we will fix the code that generated them.
For the supporting classes will still be able to receive (and will welcome)
pull requests.

Test Plan: Unit and integration tests