array_key_exists bug fix (#521)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-php-business-sdk/pull/521

This diff is associated with an external, developer-reported bug: https://developers.internmc.facebook.com/support/bugs/445172219720269/.

In PHP version 7.4, `array_key_exists()` is deprecated with Object.
`Parameters` is `extends ArrayObject`, and `array_key_exists()` is called in `Cursor` with arguments which include instance of `Parameters`. In order to avoid the error, I used `isset()` instead.

Reviewed By: jingping2015

Differential Revision: D19253757

fbshipit-source-id: 51832fbe