Fixing codegen for business SDK php [1/x]
Summary:
Due to this [change](https://www.internalfb.com/diff/D39947367?dst_version_fbid=1028137854599090&transaction_fbid=420719393500064), The PHP codegen for Business-SDK still needs to be updated. Ran codegen on the latest master so all generated code is up to date.

Need to merge this change before making any other changes to the Business SDK.

Ran this command to generate all the updated changes in the business-SDK folder
```
./scripts/codegen.sh -l php
```

Next steps: We will need to do this for the rest of the three languages. Filed a task here: T134760162

Reviewed By: ztlbells, marksliva

Differential Revision: D40335085

fbshipit-source-id: 8f8de2cc637d60b45c344682d3542854fc526811